% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a11yviz-package.R
\docType{package}
\name{a11yviz-package}
\alias{a11yviz}
\alias{a11yviz-package}
\title{a11yviz: Accessible Themes, Palettes and Audits for 'ggplot2' and 'plotly'}
\description{
Tools for making ggplot2 and plotly charts more accessible to readers
with low vision, color-vision differences, and screen-reader users.
Functions return standard ggplot or plotly objects so existing code keeps
working. The accessible CSS also styles DT tables and DiagrammeR diagrams
that appear alongside charts in Quarto and Shiny documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mshin77.github.io/a11yviz}
  \item \url{https://github.com/mshin77/a11yviz}
  \item Report bugs at \url{https://github.com/mshin77/a11yviz/issues}
}

}
\author{
\strong{Maintainer}: Mikyung Shin \email{shin.mikyung@gmail.com} (\href{https://orcid.org/0000-0001-7907-9193}{ORCID}) (affiliation: Illinois State University)

Authors:
\itemize{
  \item Mikyung Shin \email{shin.mikyung@gmail.com} (\href{https://orcid.org/0000-0001-7907-9193}{ORCID}) (affiliation: Illinois State University)
}

}
\keyword{internal}
