% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_a11y_continuous.R
\name{scale_a11y_div}
\alias{scale_a11y_div}
\alias{scale_color_a11y_div}
\alias{scale_fill_a11y_div}
\title{Accessible diverging color and fill scales}
\usage{
scale_color_a11y_div(palette = "rdbu", ...)

scale_fill_a11y_div(palette = "rdbu", ...)
}
\arguments{
\item{palette}{One of \code{"rdbu"} (default), \code{"puor"}, \code{"brbg"},
\code{"coolwarm_aaa"}.}

\item{...}{Passed to \verb{ggplot2::scale_*_gradient2()}.}
}
\value{
A ggplot2 scale.
}
\description{
ggplot2 scales for diverging gradients with anchor colors sourced from
\code{\link[=a11y_palette_div]{a11y_palette_div()}}.
}
